\name{randomLetter-package}
\alias{randomLetter-package}
\alias{randomLetter}
\docType{package}
\title{
\packageTitle{randomLetter}
}
\description{
\packageDescription{randomLetter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{randomLetter}
\packageIndices{randomLetter}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{randomLetter}

Maintainer: \packageMaintainer{randomLetter}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
